Create Source Pydantic model, add polymorphic_source to ArticleData Pydantic model
