+++ server/nlp_ssa/models/mixins/__init__.py
@@ -1,2 +1,2 @@
 from models.mixins.db import TimestampsDB, OwnedByPolymorphicSourceDB
-from models.mixins.pydantic import TimestampsMixin, OwnedByPolymorphicSourceMixin
+from models.mixins.pydantic import TimestampsMixin
