+++ server/nlp_ssa/models/source/__init__.py
@@ -1,5 +1,6 @@
 # TODO: wonder if this will eventually need to use the full module path?
 # or if other __init__'s can use the parent path instead?
-from .db import SourceDB
+from models.source.db import SourceDB
+from models.source.source import Source
 
-__all__ = ["SourceDB"]
+__all__ = ["SourceDB", "Source"]
